requirements.txt
# Core Framework
agno
pydantic            
# Auth
python-jose[cryptography]
passlib[bcrypt]
email-validator
# API Server
fastapi
uvicorn[standard]
websockets
# HTTP Client (used by agno/openai)
httpx
# Environment Variables
python-dotenv
# OpenAI (used by agno.models.openai)
openai
sqlalchemy
pytest
pytest-asyncio